refactor merge() into its own file
