+++ pandas_tutor/marks/__init__.py
@@ -1,0 +1,3 @@
+from .marks import make_marks
+
+__all__ = ["make_marks"]
